pyproject.toml
[project]
name = "artice-saver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "feedgen>=1.0.0",
    "firebase-admin>=6.8.0",
    "flask>=3.1.1",
    "python-dotenv>=1.1.0",
]
